Refactor date_picker removed _
