+++ appcui-proc-macro/src/controls.rs
@@ -24,4 +24,4 @@ pub (crate) mod numericselector;
 pub (crate) mod hline;
 pub (crate) mod vsplitter;
 pub (crate) mod vline;
-pub (crate) mod date_picker;
+pub (crate) mod datepicker;
