+++ appcui-proc-macro/src/lib.rs
@@ -452,7 +452,7 @@ pub fn vline(input: TokenStream) -> TokenStream {
 
 #[proc_macro]
 pub fn datepicker(input: TokenStream) -> TokenStream {
-    crate::controls::date_picker::create(input)
+    crate::controls::datepicker::create(input)
 }
 
 #[proc_macro]
