+++ appcui/src/prelude.rs
@@ -26,8 +26,8 @@ pub use super::ui::dropdownlist::DropDownListType;
 pub use super::ui::numericselector;
 pub use super::ui::numericselector::events::GenericNumericSelectorEvents;
 pub use super::ui::numericselector::Number;
-pub use super::ui::date_picker;
-pub use super::ui::date_picker::events::DatePickerEvents;
+pub use super::ui::datepicker;
+pub use super::ui::datepicker::events::DatePickerEvents;
 pub use super::ui::canvas;
 pub use super::ui::imageviewer;
 pub use super::ui::password;
