+++ appcui/src/ui.rs
@@ -29,7 +29,7 @@ pub mod numericselector;
 pub mod vsplitter;
 pub mod hline;
 pub mod vline;
-pub mod date_picker;
+pub mod datepicker;
 
 // re-export
 pub use common::ControlBase;
@@ -59,4 +59,4 @@ pub use numericselector::NumericSelector;
 pub use vsplitter::VSplitter;
 pub use hline::HLine;
 pub use vline::VLine;
-pub use date_picker::DatePicker;
+pub use datepicker::DatePicker;
