+++ appcui/src/ui/common/control_event_wrapper.rs
@@ -4,7 +4,7 @@ use super::traits::{Control, CustomEvents, EventProcessStatus};
 use super::UIElement;
 use crate::prelude::colorpicker::events::ColorPickerEvents;
 use crate::prelude::keyselector::events::KeySelectorEvents;
-use crate::prelude::{colorpicker, combobox, date_picker, dropdownlist, keyselector, numericselector, selector, textfield, threestatebox, GenericSelectorEvents, RuntimeManager, ThreeStateBoxEvents};
+use crate::prelude::{colorpicker, combobox, datepicker, dropdownlist, keyselector, numericselector, selector, textfield, threestatebox, GenericSelectorEvents, RuntimeManager, ThreeStateBoxEvents};
 use crate::system::Handle;
 use crate::ui::{
     button, button::events::ButtonEvents, checkbox, checkbox::events::CheckBoxEvents, password, password::events::PasswordEvents, radiobox,
@@ -13,7 +13,7 @@ use crate::ui::{
     combobox::events::ComboBoxEvents,
     dropdownlist::events::GenericDropDownListEvents,
     numericselector::events::GenericNumericSelectorEvents,
-    date_picker::events::DatePickerEvents,
+    datepicker::events::DatePickerEvents,
 };
 
 #[derive(Copy,Clone)]
@@ -36,7 +36,7 @@ pub(crate) enum ControlEventData {
     ComboBox(combobox::events::EventData),
     DropDownList(dropdownlist::events::EventData),
     NumericSelector(numericselector::events::EventData),
-    DatePicker(date_picker::events::EventData),
+    DatePicker(datepicker::events::EventData),
 }
 
 pub(crate) struct ControlEvent {
